Extract `#with_events_disabled` helper for reuse
In a subsequent commit we want to use this outside of the reporting
subscriber tests.
